+++ src/render/core/app/components/app-movie-details-header/index.css
