+++ src/render/core/app/layout/app-movie-details/index.jsx
@@ -2,11 +2,11 @@ import React from 'react'
 import all from 'it-all'
 
 import Image from 'components/app-image/'
-import BoxLoader from 'components/util-box-loader/'
-import Header from 'components/util-header/'
 import FlowText from 'components/util-flow-text/'
+import BoxLoader from 'components/util-box-loader/'
 import CustomScrollbars from 'components/util-scroller/';
 import ListCommaSplit from 'components/util-list-comma-split/'
+import AppMovieDetailsHeader from 'components/app-movie-details-header/'
 import AppMovieDetailInfo from 'components/app-movie-details-info/'
 import AppMovieDetailMenu from 'components/app-movie-details-menu/'
 
@@ -37,7 +37,7 @@ export default class MovieDetails extends React.PureComponent {
         return (
             <div className="absolute full-height movie-details">
                 {/*Close button*/}
-                <Header text="Movie" icon="icon-tv" onClick={this.props.onClick}/>
+                <AppMovieDetailsHeader text="Movie" icon="icon-tv" onClick={this.props.onClick}/>
                 <section className="row clearfix full-height margin-top-5-vh padding-left-2-vw">
                     {/*Main Loader or Movie details*/}
                     {
